Allow use of local packages without symlinks
Fleece has allowed the usage of symlinks for awhile now. This has made
it possible to reuse local Python code in multiple lambdas living in
the same codebase (such as a git repo) without needing to use packages
living on PyPi or a DevPi server.

However, the abundance of symlinks can make navigating projects
confusing to the uninitiated. Python allows for local code to be
specified using relative paths inside of a requirements.txt file, but
typically in Fleece these paths will not be available from within the
Docker container.

This adds the CLI option `--requirements-attach-directory` which lets
users specify a parent directory of the given requirements file to
attach from within docker. This makes it possible to install packages
which are defined locally. This approach will work with local tox files
and fleece as well without introducing symlinks.